agent-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
psycopg2-binary==2.9.9
pydantic==2.8.0
pydantic-settings==2.3.0
python-multipart==0.0.9
slowapi==0.1.9
